+++ extension/src/background/iocCollectionHandler.test.ts
@@ -5,7 +5,6 @@ import {
   addIocsToCollectionMessage,
   createIocCollectionMessage,
   deleteIocCollectionMessage,
-  listIocCollectionsMessage,
   removeIocFromCollectionMessage,
   renameIocCollectionMessage,
 } from "../lib/messages";